lambda-functions/online-praat-analysis/requirements.txt
boto3
numpy
scipy
librosa
soundfile
webrtcvad
soundfile
webrtcvad
matplotlib
praat-parselmouth
reportlab
